+++ cube/swiss/source/config/config.c
@@ -514,10 +514,10 @@ int config_update_global(bool checkConfigDevice) {
 	fprintf(fp, "Disable Menu Music=%s\r\n", swissSettings.disableMenuMusic ? "Yes":"No");
 	fprintf(fp, "Disable Menu SFX=%s\r\n", swissSettings.disableMenuSFX ? "Yes":"No");
 	fprintf(fp, "Menu Color=%s\r\n", uiColorStr[swissSettings.uiColor]);
-	fprintf(fp, "Library Icon=%s\r\n", faceIconStr[swissSettings.libraryIcon]);
-	fprintf(fp, "Source Icon=%s\r\n", faceIconStr[swissSettings.sourceIcon]);
-	fprintf(fp, "Settings Icon=%s\r\n", faceIconStr[swissSettings.settingsIcon]);
-	fprintf(fp, "System Icon=%s\r\n", faceIconStr[swissSettings.systemIcon]);
+	fprintf(fp, "Library Icon=%s\r\n", libraryIconStr[swissSettings.libraryIcon]);
+	fprintf(fp, "Source Icon=%s\r\n", sourceIconStr[swissSettings.sourceIcon]);
+	fprintf(fp, "Settings Icon=%s\r\n", settingsIconStr[swissSettings.settingsIcon]);
+	fprintf(fp, "System Icon=%s\r\n", systemIconStr[swissSettings.systemIcon]);
 	fprintf(fp, "Library Layout=%s\r\n", libraryLayoutStr[swissSettings.libraryLayout]);
 	fprintf(fp, "Init DVD Drive at startup=%s\r\n", swissSettings.initDVDDriveAtStart ? "Yes":"No");
 	fprintf(fp, "Stop DVD Drive motor=%s\r\n", swissSettings.stopMotor ? "Yes":"No");
@@ -1339,32 +1339,32 @@ void config_parse_global(char *configData) {
 					}
 				}
 				else if(!strcmp("Library Icon", name)) {
-					for(int i = 0; i < UI_HOME_ICON_COUNT; i++) {
-						if(!strcmp(faceIconStr[i], value)) {
+					for(int i = 0; i < UI_HOME_ICON_CHOICES; i++) {
+						if(!strcmp(libraryIconStr[i], value)) {
 							swissSettings.libraryIcon = i;
 							break;
 						}
 					}
 				}
 				else if(!strcmp("Source Icon", name)) {
-					for(int i = 0; i < UI_HOME_ICON_COUNT; i++) {
-						if(!strcmp(faceIconStr[i], value)) {
+					for(int i = 0; i < UI_HOME_ICON_CHOICES; i++) {
+						if(!strcmp(sourceIconStr[i], value)) {
 							swissSettings.sourceIcon = i;
 							break;
 						}
 					}
 				}
 				else if(!strcmp("Settings Icon", name)) {
-					for(int i = 0; i < UI_HOME_ICON_COUNT; i++) {
-						if(!strcmp(faceIconStr[i], value)) {
+					for(int i = 0; i < UI_HOME_ICON_CHOICES; i++) {
+						if(!strcmp(settingsIconStr[i], value)) {
 							swissSettings.settingsIcon = i;
 							break;
 						}
 					}
 				}
 				else if(!strcmp("System Icon", name)) {
-					for(int i = 0; i < UI_HOME_ICON_COUNT; i++) {
-						if(!strcmp(faceIconStr[i], value)) {
+					for(int i = 0; i < UI_HOME_ICON_CHOICES; i++) {
+						if(!strcmp(systemIconStr[i], value)) {
 							swissSettings.systemIcon = i;
 							break;
 						}